Unset GEM_HOME in zshrc
This is due to the tmuxinator tool setting GEM_HOME when installed via
homebrew. It has to set this so that it knows where to find its
dependencies no matter what the current ruby version is. Normally this
is fine, but since tmuxinator starts new shells, this var is inherited
and wreaks havoc.